api/requirements.txt
azure-functions>=1.17.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-eventgrid>=4.17.0
httpx>=0.26.0
aiohttp>=3.9.0
pydantic>=2.5.0
